#112 new component for preview
